Refactor I2SAudioMicrophone read method to use uint8_t buffer and simplify setup

+++ components/m5atoms3_audio/microphone/i2s_audio_microphone.cpp
@@ -16,25 +16,6 @@ static const size_t BUFFER_SIZE = 512;
 static const char *const TAG = "m5atoms3.microphone";
 
 void I2SAudioMicrophone::setup() {
-//   ESP_LOGCONFIG(TAG, "Setting up I2S Audio Microphone...");
-// #if SOC_I2S_SUPPORTS_ADC
-//   if (this->adc_) {
-//     if (this->parent_->get_port() != I2S_NUM_0) {
-//       ESP_LOGE(TAG, "Internal ADC only works on I2S0!");
-//       this->mark_failed();
-//       return;
-//     }
-//   } else
-// #endif
-//       if (this->pdm_) {
-//     if (this->parent_->get_port() != I2S_NUM_0) {
-//       ESP_LOGE(TAG, "PDM only works on I2S0!");
-//       this->mark_failed();
-//       return;
-//     }
-//   }
-
-
 
   ESP_LOGI(TAG, "setup");
 
@@ -136,58 +117,8 @@ void I2SAudioMicrophone::stop_() {
   this->high_freq_.stop();
 }
 
-size_t I2SAudioMicrophone::read(int16_t *buf, size_t len) {
-  // size_t bytes_read = 0;
-  // esp_err_t err = i2s_read(this->parent_->get_port(), buf, len, &bytes_read, (100 / portTICK_PERIOD_MS));
-  // if (err != ESP_OK) {
-  //   ESP_LOGW(TAG, "Error reading from I2S microphone: %s", esp_err_to_name(err));
-  //   this->status_set_warning();
-  //   return 0;
-  // }
-  // if (bytes_read == 0) {
-  //   this->status_set_warning();
-  //   return 0;
-  // }
-  // this->status_clear_warning();
-  // if (this->bits_per_sample_ == I2S_BITS_PER_SAMPLE_16BIT) {
-  //   return bytes_read;
-  // } else if (this->bits_per_sample_ == I2S_BITS_PER_SAMPLE_32BIT) {
-  //   std::vector<int16_t> samples;
-  //   size_t samples_read = bytes_read / sizeof(int32_t);
-  //   samples.resize(samples_read);
-  //   for (size_t i = 0; i < samples_read; i++) {
-  //     int32_t temp = reinterpret_cast<int32_t *>(buf)[i] >> 14;
-  //     samples[i] = clamp<int16_t>(temp, INT16_MIN, INT16_MAX);
-  //   }
-  //   memcpy(buf, samples.data(), samples_read * sizeof(int16_t));
-  //   return samples_read * sizeof(int16_t);
-  // } else {
-  //   ESP_LOGE(TAG, "Unsupported bits per sample: %d", this->bits_per_sample_);
-  //   return 0;
-  // }
-
-
-  // size_t bytes_read = 0;
-  // esp_err_t err = i2s_read(this->parent_->get_port(), buf, len, &bytes_read, (100 / portTICK_PERIOD_MS));
-  // if (err != ESP_OK) {
-  //   ESP_LOGW(TAG, "Error reading from I2S microphone: %s", esp_err_to_name(err));
-  //   this->status_set_warning();
-  //   return 0;
-  // }
-  // if (bytes_read == 0) {
-  //   this->status_set_warning();
-  //   return 0;
-  // }
-
-
-
-
-
-  // // Record into buffer 
-  // ESP_LOGI(TAG, "rec %d", BUFFER_SIZE);
-
-
-  // M5.Mic.record(buf, len >> 1, 16000);
+size_t I2SAudioMicrophone::read(uint8_t *buf, size_t len) {
+  
   M5.Mic.record(buf, 256, 16000);
   // M5.Mic.record(buf, 512, 16000);
   while (M5.Mic.isRecording());
